Refactor: add helper to format date
