+++ src/helpers/FormatDate.js
@@ -1,0 +1,8 @@
+export const formatDate = (date) => {
+	const newDate = new Date(date);
+	const options = {
+		month: 'long',
+		day: '2-digit',
+	};
+	return new Intl.DateTimeFormat('es', options).format(newDate);
+};
